Add more hooks to be able to detect keyboard/mouse activity.
I've added a low-level mouse hook and callbacks for general mouse and
keyboard events.  I wanted to be able to hide the command feedback message
window without having to hunt down the escape key and the like.  I've
changed the code to use these events to automatically hide the message when
a key is pressed or something happens with the mouse.

I've also cleaned up the code with regards to the above changes and added
an 'eval' command that passes its arguments to Python's eval().