Add a convert command.
I thought that it would be useful to have some sort of ability to convert
values with given units into other compatible units.  I already have a
conversion package and thus it was easy to add this command in.  It simply
takes a value, its unit and the desired unit and then displays the
resulting conversion value if all's well.