Pass unknown commands to eval command if present.
Rather than having to type 'eval' all of the time, pass unknown commands to
the eval command if an eval command is present.  Having a hook allows one
to code their own version rather than having a hard-coded one.  If no eval
command is available, the code bitches as usual.