Make it so that input intended for us doesn't go somewhere else.
Since I'm trying to be as generic as possible, I'm depending on the window
manager to set the proper focus, etc.  Needless to say, this doesn't always
work out.  So, I've added code to the keyboard hook that checks to see if
we have the focus.  If not, it doesn't let the keyboard input through.
This way it'll be obvious if we're missing the target, input-wise.  (This
code isn't as portable as I would like it to be.  It depends on
Windows-specific stuff.)