Add clipboard-access command and enhance EvalCommand with clipboard access.
The eval command now tries to get its argument from the clipboard if no
argument is given via the command input.  A successful eval() now puts the
result on the clipboard as well.  The clipboard access allows one to see
and set the contents of the clipboard.  (Only text is supported.)