Add tab completion.
Well, with a little help from reading the Qt source code via the gitorious
repository, I've finally been able to add tab completion and have it work
the way that I wanted it to.  I've added a Completer class that
re-implements eventFilter() so that the tab and backtabs aren't filtered
out any more.  Now, the interaction with the popup using tab and backtab
behaves the same way as using the down and up arrows, respectively.