Add argument-required decorator and clean up commands.
In order to be better able to give feedback to the user, I've changed the
commands so that they raise exceptions if there is a problem instead of
detecting the errors within the command implementation.  This way, the
command manager can detect the errors and indicate problems to the user and
then indicate to the input manager which command inputs can go into the
history.  (The infrastructure was already in place but I wasn't
consistently using it.)  I've also added a decorator to simplify commands
that require an argument.